+++ api/helpers/rest/get-queue.js
@@ -1,4 +1,4 @@
-/* global sails, Meta, _, needle */
+/* global sails, Meta, _, needle, Status */
 
 module.exports = {
 
@@ -13,23 +13,27 @@ module.exports = {
     fn: async function (inputs, exits) {
         sails.log.debug('Helper rest.getQueue called.');
         try {
+            // Return queue in memory instead of checking for the current queue if we are waiting for a healthy RadioDJ to report
+            if (Status.errorCheck.waitForGoodRadioDJ)
+                return exits.success(Meta.automation);
             // Query for the radioDJ queue and update Meta.automation with the queue.
             needle('get', Meta['A'].radiodj + '/p?auth=' + sails.config.custom.rest.auth, {}, {open_timeout: 2000, response_timeout: 2000, read_timeout: 2000, headers: {'Content-Type': 'application/json'}})
                     .then(async function (resp) {
                         try {
-                            Meta.automation = [];
                             if (typeof resp.body.name === 'undefined' || (resp.body.name !== 'ArrayOfSongData' && resp.body.name !== 'SongData'))
                             {
                                 return exits.success([]);
                             }
                             if (resp.body.name === 'ArrayOfSongData')
                             {
+                                Meta.automation = [];
                                 resp.body.children.map(trackA => {
                                     var theTrack = {};
                                     trackA.children.map(track => theTrack[track.name] = track.value);
                                     Meta.automation.push(theTrack);
                                 });
                             } else {
+                                Meta.automation = [];
                                 var theTrack = {};
                                 resp.body.children.map(track => theTrack[track.name] = track.value);
                                 Meta.automation.push(theTrack);