+++ api/models/Status.js
@@ -132,9 +132,10 @@ module.exports = {
                         await sails.helpers.rest.cmd('EnableAssisted', 1, 0);
                         await sails.helpers.rest.cmd('EnableAutoDJ', 1, 0);
                         await sails.helpers.rest.cmd('StopPlayer', 1, 0);
+                        var queue = Meta.automation;
                         await sails.helpers.rest.changeRadioDj();
                         await sails.helpers.rest.cmd('ClearPlaylist', 1);
-                        await sails.helpers.error.post();
+                        await sails.helpers.error.post(queue);
                         await Meta.changeMeta({changingState: null});
                         return resolve(0);
                     } catch (e) {
@@ -190,9 +191,10 @@ module.exports = {
                             await sails.helpers.rest.cmd('EnableAutoDJ', 0, 0);
                             await sails.helpers.rest.cmd('EnableAssisted', 1, 0);
                             await sails.helpers.rest.cmd('StopPlayer', 0, 0);
+                            var queue = Meta.automation;
                             await sails.helpers.rest.changeRadioDj();
                             await sails.helpers.rest.cmd('ClearPlaylist', 1);
-                            await sails.helpers.error.post();
+                            await sails.helpers.error.post(queue);
                             await Meta.changeMeta({changingState: null});
                         }
                         return resolve(0);
@@ -254,9 +256,10 @@ module.exports = {
                             await sails.helpers.rest.cmd('EnableAutoDJ', 0, 0);
                             await sails.helpers.rest.cmd('EnableAssisted', 1, 0);
                             await sails.helpers.rest.cmd('StopPlayer', 0, 0);
+                            var queue = Meta.automation;
                             await sails.helpers.rest.changeRadioDj();
                             await sails.helpers.rest.cmd('ClearPlaylist', 1);
-                            await sails.helpers.error.post();
+                            await sails.helpers.error.post(queue);
                             await Meta.changeMeta({changingState: null});
                         }
                         return resolve(0);